Move Task-related classes to main.java.tasks
Moved Task, TaskList, Todo, Deadline and Event classes into
main.java.tasks package.
