+++ src/main/java/Karen.java
@@ -1,11 +1,5 @@
-import java.time.format.DateTimeParseException;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Scanner;
-
-import main.java.*;
-
 import main.java.commands.Command;
+import main.java.tasks.TaskList;
 import main.java.util.Parser;
 import main.java.util.Storage;
 import main.java.util.Ui;
